AI Agents/Nutrition Consultant/pyproject.toml
[tool.poetry]
name = "nutrition consultant"
version = "0.1.0"
description = ""
authors = ["caiocmb7 <caiobarros@alu.ufc.br>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<=3.13"
crewai = "^0.51.1"
crewai-tools = "^0.8.3"
langchain = "^0.2.14"
langchain-openai = "^0.1.21"
python-dotenv = "^1.0.1"
langchain-groq = "^0.1.9"
jupyterlab = "^4.2.4"
black = "^24.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
